requirements.txt
django
pymongo
dnspython
django-cors-headers
python-dotenv
gunicorn